Create README.md
Thousands of houses are sold everyday. There are some questions every buyer asks himself like: What is the actual price that this house deserves? Am I paying a fair price? The real estate sector is an important industry with many stakeholders ranging from regulatory bodies to private companies and investors.
The aim of the project is to build a predictive model for house price based on various feature.The price of the house depends on many features and varies accordingly